+++ src/main/java/com/google/cloud/tools/gradle/appengine/core/DeployExtension.java
@@ -45,11 +45,12 @@ public class DeployExtension
     this.gradleProject = gradleProject;
   }
 
-  /** Copy constructor for DeployExtension. */
+  /** Creates and return a deep copy of the DeployExtension. */
   public DeployExtension(DeployExtension deployExtension) {
     this.gradleProject = deployExtension.gradleProject;
     this.bucket = deployExtension.bucket;
-    this.deployables = new ArrayList<>(deployExtension.deployables);
+    this.deployables =
+        deployExtension.deployables == null ? null : new ArrayList<>(deployExtension.deployables);
     this.imageUrl = deployExtension.imageUrl;
     this.project = deployExtension.project;
     this.promote = deployExtension.promote;